% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{charitable}
\alias{charitable}
\title{charitable data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/Charitable}
}
\usage{
charitable
}
\description{
Data used for the paper "Does Price matter in charitable giving? Evidence
from a large-Scale Natural Field experiment" by Karlan and  List (2007).
}
\keyword{datasets}
