% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDetails.R
\name{dataDetails}
\alias{dataDetails}
\title{Opens a link to the paper behind a given method}
\usage{
dataDetails(datasetName = NULL)
}
\arguments{
\item{datasetName}{A string with the name of the dataset}
}
\description{
\code{dataDetails} opens the browser and displays the paper that uses the
  dataset
}
\examples{
dataDetails("social")

}
