% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{IVdataset}
\alias{IVdataset}
\title{IV data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/IV\%20Datasets}
}
\usage{
IVdataset
}
\description{
Data used for the paper "Does Compulsory School Attendance Affect Schooling
and Earnings?" Angrist and Kruger (1991) and related papers.
}
\keyword{datasets}
