% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{mobilization}
\alias{mobilization}
\title{mobilization data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/Mobilization}
}
\usage{
mobilization
}
\description{
Data used for the paper "Comparing Experimental and Matching Methods Using a
Large-Scale Voter Mobilization Experiment" Arceneaux, Gerber and Green (2006).
}
\keyword{datasets}
