% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{secrecy}
\alias{secrecy}
\title{secrecy data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/Secrecy}
}
\usage{
secrecy
}
\description{
Data used for the paper "Ballot Secrecy Concerns and Voter Mobilization: New
Experimental Evidence about Message Source, Context, and the Duration of
Mobilization Effects" by Gerber, Hubers, Biggers, and Hendry (2014).
}
\keyword{datasets}
