% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{social}
\alias{social}
\title{social data}
\format{A tibble with variables:
\describe{
\item{outcome_voted}{Response variable: Indicator variable where =1 indicates voted in the August 2006 primary}
\item{treat_neighbors}{Treatment: Indicator variable where =1 indicates Neighbors mailing treatment Covariates}
\item{sex}{Indicator variable where =1 indicates male}
\item{yob}{Year of birth}
\item{g2000}{Indicator variable where =1 indicates voted in the 2000 general}
\item{g2002}{Indicator variable where =1 indicates voted in the 2002 general}
\item{p2000}{Indicator variable where =1 indicates voted in the 2000 primary}
\item{p2002}{Indicator variable where =1 indicates voted in the 2002 primary}
\item{p2004}{Indicator variable where =1 indicates voted in the 2004 primary}
\item{city}{City index}
\item{hh_size}{Household size}
\item{totalpopulation_estimate}{Estimate of city population}
\item{percent_male}{Percentage males in household}
\item{median_age}{Median age in household}
\item{median_income}{Median income in household}
\item{percent_62yearsandover}{Percentage of subjects of age higher than 62 yo}
\item{percent_white}{Percentage white in household}
\item{percent_black}{Percentage black in household}
\item{percent_asian}{Percentage asian in household}
\item{percent_hispanicorlatino}{Percentage hispanic or latino in household}
\item{employ_20to64}{Percentage of employed subjects of age 20 to 64 yo}
\item{highschool}{Percentage having only high school degree}
\item{bach_orhigher}{Percentage having bachelor degree or higher}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/Social}
}
\usage{
social
}
\description{
Data used for the paper "Social Pressure and Voter Turnout: Evidence from a
Large-Scale Field Experiment" by Gerber, Green, and Larimer (2008).
}
\keyword{datasets}
