% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{vouchers}
\alias{vouchers}
\title{school vouchers data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/School\%20Vouchers}
}
\usage{
vouchers
}
\description{
Data used for the paper "Vouchers for Private Schooling in Colombia: Evidence
from a Randomized Natural Experiment" Angrist, Bettinger, Bloom, King, and
Kremer (2002).
}
\keyword{datasets}
