% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentData.R
\docType{data}
\name{welfare}
\alias{welfare}
\title{welfare data}
\format{A tibble with variables:
\describe{
\item{TBA}{TBA}
}}
\source{
\url{https://github.com/gsbDBI/ExperimentData/tree/master/Welfare}
}
\usage{
welfare
}
\description{
Data used for the paper "Modeling heterogeneous treatment effects in survey
experiments with Bayesian Additive Regression Trees" Green, Kern (2012)
}
\keyword{datasets}
